+++ trunk/test/hdf5lib/TestH5Edefault.java
@@ -51,8 +51,8 @@ public class TestH5Edefault {
     public void testH5Eget_current_stack() {
         long num_msg = -1;
         long num_msg_default = -1;
-        int stack_id = -1;
-        int stack_id_default = HDF5Constants.H5E_DEFAULT;
+        long stack_id = -1;
+        long stack_id_default = HDF5Constants.H5E_DEFAULT;
         try {
             H5.H5Fopen("test", 0, 1); 
         }
@@ -120,7 +120,7 @@ public class TestH5Edefault {
     public void testH5Eget_current_stack_pop() {
         long num_msg = -1;
         long num_msg_default = -1;
-        int stack_id = -1;
+        long stack_id = -1;
         try {
             H5.H5Fopen("test", 0, 1);
         }
@@ -250,7 +250,7 @@ public class TestH5Edefault {
     @Test
     public void testH5Ecreate_stack() {
         try {
-            int stack_id = H5.H5Ecreate_stack();
+            long stack_id = H5.H5Ecreate_stack();
             assertTrue("H5.H5Ecreate_stack", stack_id > 0);
             H5.H5Eclose_stack(stack_id);
         }
@@ -268,7 +268,7 @@ public class TestH5Edefault {
     @Test
     public void testH5Eset_current_stack() {
         long num_msg = -1;
-        int stack_id = -1;
+        long stack_id = -1;
         try {
             H5.H5Fopen("test", 0, 1);
         }
